pyproject.toml
[tool.poetry]
name = "claude_otel_session_scorer"
version = "0.1.0"
description = "A series of data pipelines for Databricks to score claude code sessions collected from Open Telemetry"
authors = ["Tanner Wendland"]
readme = "README.md"
packages = [{include = "claude_otel_session_scorer"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
databricks-connect = "==18.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
ruff = ">=0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
claude_otel_session_scorer = "claude_otel_session_scorer.main:main"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
